requirements.txt
beautifulsoup4==4.9.3
certifi==2021.10.8
chardet==4.0.0
googlesearch-python==1.1.0
idna==2.10
requests==2.25.1
soupsieve==2.3.1
urllib3==1.26.8
